+++ lib/scala/docs-generator/src/main/scala/org/enso/DocsGeneratorMain.scala
@@ -4,105 +4,12 @@ import java.io._
 
 import scala.util.Using
 import scala.io.Source
-import org.enso.syntax.text.{DocParser, DocParserMain, Parser}
 import scalatags.Text.{all => HTML}
 import HTML._
 
-object TreeOfCommonPrefixes {
-  case class Node(name: String, var elems: List[Node]) {
-    def html(beg: String = ""): Modifier = {
-      val newBeg = if (beg.length > 0) beg + "-" + name else name
-      if (elems.isEmpty) {
-        HTML.li(HTML.a(HTML.href := newBeg)(name))
-      } else {
-        HTML.li(HTML.`class` := "section")(
-          HTML.input(HTML.`type` := "checkbox", HTML.id := newBeg),
-          HTML.label(HTML.`for` := newBeg)(
-            HTML.a(HTML.href := newBeg)(name)
-          ),
-          elems.map(x => HTML.ul(x.html(newBeg)))
-        )
-      }
-    }
-  }
-
-  def groupNodesByPrefix(le: List[Node]): List[Node] =
-    groupByPrefix(le.map(_.name))
-
-  def groupByPrefix(ls: List[String]): List[Node] = {
-    var nodes = List[Node]()
-    for (string <- ls) {
-      if (string.split('/').length <= 1) {
-        nodes = nodes :+ Node(string, List())
-      } else {
-        val arr      = string.split('/')
-        val filtered = nodes.filter(x => x.name == arr.head)
-        if (filtered.nonEmpty && nodes.contains(filtered.head)) {
-          nodes.map(n =>
-            if (n == filtered.head) {
-              n.elems = n.elems :+ Node(arr.tail.mkString("/"), List())
-            }
-          )
-        } else {
-          nodes =
-            nodes :+ Node(arr.head, List(Node(arr.tail.mkString("/"), List())))
-        }
-      }
-    }
-    for (node <- nodes) {
-      node.elems = groupNodesByPrefix(node.elems)
-    }
-    nodes
-  }
-}
-
-object DocsGenerator {
-
-  def generate(program: String): String = {
-    val parser   = new Parser()
-    val module   = parser.run(program)
-    val dropMeta = parser.dropMacroMeta(module)
-    val doc      = DocParser.DocParserRunner.createDocs(dropMeta)
-    val code =
-      DocParser.DocParserHTMLGenerator.generateHTMLForEveryDocumented(doc)
-    code
-  }
-
-  def generatePure(comment: String): String = {
-    val doc  = DocParserMain.runMatched(comment)
-    val html = DocParser.DocParserHTMLGenerator.generateHTMLPureDoc(doc)
-    html
-  }
-
-  def mapIfEmpty(doc: String): String = {
-    var tmp = doc
-    if (doc.replace("<div>", "").replace("</div>", "").length == 0) {
-      tmp =
-        "\n\n*Enso Reference Viewer.*\n\nNo documentation available for chosen source file."
-      tmp = generatePure(tmp).replace("style=\"font-size: 13px;\"", "")
-    }
-    tmp
-  }
-
-  def removeUnnecessaryDivs(doc: String): String = {
-    var tmp = doc
-    while (tmp.contains("<div></div>"))
-      tmp = tmp.replace("<div></div>", "")
-    tmp
-  }
-
-  def traverse(root: File): LazyList[File] =
-    if (!root.exists) LazyList.empty
-    else
-      LazyList.apply(root) ++ (root.listFiles match {
-        case null  => LazyList.empty
-        case files => files.view.flatMap(traverse)
-      })
-}
-
 object DocsGeneratorMain extends App {
-  import DocsGenerator._
-  import TreeOfCommonPrefixes._
+  import docsgenerator.DocsGenerator._
+  import docsgenerator.TreeOfCommonPrefixes._
 
   /// Files
   val path = "./distribution/std-lib/Standard/src"
@@ -168,9 +75,7 @@ object DocsGeneratorMain extends App {
     val bw = new BufferedWriter(new FileWriter(file))
     var treeCode =
       "<div>" + treeStyle + HTML
-        .ul(
-          treeNames.map(_.html())
-        )
+        .ul(treeNames.map(_.html()))
         .render
         .replace("class=", "className=") + "</div>"
     if (x._1.contains("/index.js")) {
